feat: tweaks to chain fields and reconciling types.

+++ postgres-driver/load_balancer.go
@@ -25,8 +25,8 @@ type dbLoadBalancer struct {
 	UserID            sql.NullString `db:"user_id"`
 	AppIDS            sql.NullString `db:"app_ids"`
 	Origins           pq.StringArray `db:"origins"`
-	RelaysLimit       sql.NullInt64  `db:"relays_limit"`
-	RequestTimeout    sql.NullInt64  `db:"request_timeout"`
+	RelaysLimit       sql.NullInt32  `db:"relays_limit"`
+	RequestTimeout    sql.NullInt32  `db:"request_timeout"`
 	Gigastake         sql.NullBool   `db:"gigastake"`
 	GigastakeRedirect sql.NullBool   `db:"gigastake_redirect"`
 	Stickiness        sql.NullBool   `db:"stickiness"`
@@ -41,13 +41,13 @@ func (lb *dbLoadBalancer) toLoadBalancer() *repository.LoadBalancer {
 		Name:              lb.Name.String,
 		UserID:            lb.UserID.String,
 		ApplicationIDs:    strings.Split(lb.AppIDS.String, ","),
-		RequestTimeout:    lb.RequestTimeout.Int64,
+		RequestTimeout:    int(lb.RequestTimeout.Int32),
 		Gigastake:         lb.Gigastake.Bool,
 		GigastakeRedirect: lb.GigastakeRedirect.Bool,
 		StickyOptions: repository.StickyOptions{
 			Duration:      lb.Duration.String,
 			StickyOrigins: lb.Origins,
-			RelaysLimit:   int(lb.RelaysLimit.Int64),
+			RelaysLimit:   int(lb.RelaysLimit.Int32),
 			Stickiness:    lb.Stickiness.Bool,
 			UseRPCID:      lb.UseRPCID.Bool,
 		},
